All types and enums extraxted and helper functions added to utils file
