+++ client-state/Form.ts
@@ -1,9 +1,9 @@
 import Joi from 'joi'
-import { VisibleFieldsObject } from '../components/EligibilityPage'
 import { getWebTranslations, WebTranslations } from '../i18n/web'
 import { BenefitHandler } from '../utils/api/benefitHandler'
 import { Language, ValidationErrors } from '../utils/api/definitions/enums'
 import { FieldConfig, FieldKey } from '../utils/api/definitions/fields'
+import { VisibleFieldsObject } from '../utils/api/definitions/types'
 import MainHandler from '../utils/api/mainHandler'
 import { FormField } from './FormField'
 import { InputHelper } from './InputHelper'
